update paging for promotions and remove decimal places

+++ src/main/java/com/aplusplus/HotelBooking/dto/PromotionDTO.java
@@ -13,7 +13,7 @@ import java.util.HashSet;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class PromotionDTO {
     private Long id;
-    private Double percentOfDiscount;
+    private Integer percentOfDiscount;
     private String promotionTitle;
     private String description;
     private LocalDate startDate;
